+++ integrationtests/scenarios/i002528-5/before/Library.fs
@@ -1,0 +1,5 @@
+namespace before
+
+module Say =
+    let hello name =
+        printfn "Hello %s" name